Loop autônomo (residuais): turnover cast Decimal->Integer + profile/tipos-posto 422
- turnover_repository: func.cast(bool_col, Decimal) usava o Decimal do Python (nao é
  tipo SQLAlchemy) -> _static_cache_key 500. Trocado p/ Integer (bool->int; NUMERIC
  não coage boolean no Postgres). Corrige o erro do cast; 4 casts.
- profile_controller: /tipos-posto era sombreado por /{profile_id} (str) -> uuid cast
  500. profile_id str->UUID. Provado 500->422.

RESIDUAL turnover/dashboard (ainda 500, SEPARADO): get_aggregated_factors usa
count(distinct(scalar_subquery correlacionado)) -> GROUP BY error no Postgres. Precisa
rewrite p/ JOIN (funcionarios_afetados). Modulo recém-onlined (sem users) — baixa urgência.

+++ backend/modules/retention/turnover/repositories/turnover_repository.py
@@ -11,7 +11,7 @@ from decimal import Decimal
 from typing import Any
 from uuid import UUID
 
-from sqlalchemy import and_, delete, func, or_, select, update
+from sqlalchemy import Integer, and_, delete, func, or_, select, update
 from sqlalchemy import inspect as sa_inspect
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
@@ -340,7 +340,7 @@ class TurnoverRepository:
                 func.count(RiskFactor.id).label("total_ocorrencias"),
                 func.avg(RiskFactor.contribuicao_score).label("contribuicao_media"),
                 func.max(RiskFactor.contribuicao_score).label("contribuicao_maxima"),
-                func.avg(func.cast(RiskFactor.threshold_violado, Decimal)).label("percentual_threshold"),
+                func.avg(func.cast(RiskFactor.threshold_violado, Integer)).label("percentual_threshold"),
                 func.count(
                     func.distinct(
                         select(TurnoverPrediction.funcionario_id)
@@ -574,13 +574,13 @@ class TurnoverRepository:
                 func.sum(
                     func.cast(
                         TurnoverPrediction.nivel == NivelRisco.CRITICO,
-                        Decimal,
+                        Integer,
                     )
                 ).label("total_criticos"),
                 func.sum(
                     func.cast(
                         TurnoverPrediction.nivel == NivelRisco.ALTO,
-                        Decimal,
+                        Integer,
                     )
                 ).label("total_altos"),
             )
